chore: extract license plugin
