+++ package.json
@@ -67,6 +67,7 @@
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
     "execa": "^6.1.0",
+    "fast-glob": "^3.2.11",
     "fs-extra": "^10.1.0",
     "lint-staged": "^13.0.3",
     "minimist": "^1.2.6",
@@ -76,8 +77,10 @@
     "pnpm": "^7.9.3",
     "prettier": "2.7.1",
     "prompts": "^2.4.2",
+    "resolve": "^1.22.1",
     "rimraf": "^3.0.2",
     "rollup": "~2.78.0",
+    "rollup-plugin-license": "^2.8.1",
     "semver": "^7.3.7",
     "simple-git-hooks": "^2.8.0",
     "tslib": "^2.4.0",
