+++ packages/vite/rollup.config.ts
@@ -1,17 +1,13 @@
 /* eslint-disable no-restricted-globals */
-import fs from 'node:fs'
 import path from 'node:path'
 import nodeResolve from '@rollup/plugin-node-resolve'
 import typescript from '@rollup/plugin-typescript'
 import commonjs from '@rollup/plugin-commonjs'
 import json from '@rollup/plugin-json'
-import license from 'rollup-plugin-license'
 import MagicString from 'magic-string'
-import colors from 'picocolors'
-import fg from 'fast-glob'
-import { sync as resolve } from 'resolve'
 import type { Plugin, RollupOptions } from 'rollup'
 import { defineConfig } from 'rollup'
+import licensePlugin from '../../scripts/rollupLicensePlugin.mjs'
 import pkg from './package.json'
 
 const envConfig = defineConfig({
@@ -254,115 +250,6 @@ function shimDepsPlugin(deps: Record<string, ShimOptions>): Plugin {
       }
     }
   }
-}
-
-function licensePlugin() {
-  return license({
-    thirdParty(dependencies) {
-      // https://github.com/rollup/rollup/blob/master/build-plugins/generate-license-file.js
-      // MIT Licensed https://github.com/rollup/rollup/blob/master/LICENSE-CORE.md
-      const coreLicense = fs.readFileSync(
-        path.resolve(__dirname, '../../LICENSE')
-      )
-      function sortLicenses(licenses) {
-        let withParenthesis = []
-        let noParenthesis = []
-        licenses.forEach((license) => {
-          if (/^\(/.test(license)) {
-            withParenthesis.push(license)
-          } else {
-            noParenthesis.push(license)
-          }
-        })
-        withParenthesis = withParenthesis.sort()
-        noParenthesis = noParenthesis.sort()
-        return [...noParenthesis, ...withParenthesis]
-      }
-      const licenses = new Set()
-      const dependencyLicenseTexts = dependencies
-        .sort(({ name: nameA }, { name: nameB }) =>
-          nameA > nameB ? 1 : nameB > nameA ? -1 : 0
-        )
-        .map(
-          ({
-            name,
-            license,
-            licenseText,
-            author,
-            maintainers,
-            contributors,
-            repository
-          }) => {
-            let text = `## ${name}\n`
-            if (license) {
-              text += `License: ${license}\n`
-            }
-            const names = new Set()
-            for (const person of [author, ...maintainers, ...contributors]) {
-              const name = typeof person === 'string' ? person : person?.name
-              if (name) {
-                names.add(name)
-              }
-            }
-            if (names.size > 0) {
-              text += `By: ${Array.from(names).join(', ')}\n`
-            }
-            if (repository) {
-              text += `Repository: ${
-                typeof repository === 'string' ? repository : repository.url
-              }\n`
-            }
-            if (!licenseText) {
-              try {
-                const pkgDir = path.dirname(
-                  resolve(path.join(name, 'package.json'), {
-                    preserveSymlinks: false
-                  })
-                )
-                const licenseFile = fg.sync(`${pkgDir}/LICENSE*`, {
-                  caseSensitiveMatch: false
-                })[0]
-                if (licenseFile) {
-                  licenseText = fs.readFileSync(licenseFile, 'utf-8')
-                }
-              } catch {}
-            }
-            if (licenseText) {
-              text +=
-                '\n' +
-                licenseText
-                  .trim()
-                  .replace(/(\r\n|\r)/gm, '\n')
-                  .split('\n')
-                  .map((line) => `> ${line}`)
-                  .join('\n') +
-                '\n'
-            }
-            licenses.add(license)
-            return text
-          }
-        )
-        .join('\n---------------------------------------\n\n')
-      const licenseText =
-        `# Vite core license\n` +
-        `Vite is released under the MIT license:\n\n` +
-        coreLicense +
-        `\n# Licenses of bundled dependencies\n` +
-        `The published Vite artifact additionally contains code with the following licenses:\n` +
-        `${sortLicenses(licenses).join(', ')}\n\n` +
-        `# Bundled dependencies:\n` +
-        dependencyLicenseTexts
-      const existingLicenseText = fs.readFileSync('LICENSE.md', 'utf8')
-      if (existingLicenseText !== licenseText) {
-        fs.writeFileSync('LICENSE.md', licenseText)
-        console.warn(
-          colors.yellow(
-            '\nLICENSE.md updated. You should commit the updated file.\n'
-          )
-        )
-      }
-    }
-  })
 }
 
 /**
